User en Lift aangepast zodat ze random doubles accepteren

+++ Liften/src/Model/User.java
@@ -27,9 +27,9 @@ public class User {
     @JsonProperty("arrivalTime")
     private Double arrivalTime;
     @JsonProperty("boardingTime")
-    private Integer boardingTime;
+    private Double boardingTime;
     @JsonProperty("unboardingTime")
-    private Integer unboardingTime;
+    private Double unboardingTime;
     @JsonProperty("timeout")
     private Integer timeout;
     @JsonProperty("source-id")
@@ -111,7 +111,7 @@ public class User {
      *     The boardingTime
      */
     @JsonProperty("boardingTime")
-    public Integer getBoardingTime() {
+    public Double getBoardingTime() {
         return boardingTime;
     }
 
@@ -121,7 +121,7 @@ public class User {
      *     The boardingTime
      */
     @JsonProperty("boardingTime")
-    public void setBoardingTime(Integer boardingTime) {
+    public void setBoardingTime(Double boardingTime) {
         this.boardingTime = boardingTime;
     }
 
@@ -131,7 +131,7 @@ public class User {
      *     The unboardingTime
      */
     @JsonProperty("unboardingTime")
-    public Integer getUnboardingTime() {
+    public Double getUnboardingTime() {
         return unboardingTime;
     }
 
@@ -141,7 +141,7 @@ public class User {
      *     The unboardingTime
      */
     @JsonProperty("unboardingTime")
-    public void setUnboardingTime(Integer unboardingTime) {
+    public void setUnboardingTime(Double unboardingTime) {
         this.unboardingTime = unboardingTime;
     }
